+++ src/app/(layout-with-navigation)/(pages-statiques)/nouveautes/[slug]/page.tsx
@@ -1,5 +1,5 @@
 import Route404 from '@/components/layout/404'
-import Meta from '@/components/misc/Meta'
+import Meta from '@/components/misc/MetaOpenGraph'
 import InlineLink from '@/design-system/inputs/InlineLink'
 import Title from '@/design-system/layout/Title'
 import Markdown from '@/design-system/utils/Markdown'
