+++ src/app/(layout-with-navigation)/(simulation)/actions/plus/[...dottedName]/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import Meta from '@/components/misc/Meta'
+import Meta from '@/components/misc/MetaOpenGraph'
 import Trans from '@/components/translation/Trans'
 import ButtonLink from '@/design-system/inputs/ButtonLink'
 import Markdown from '@/design-system/utils/Markdown'
