+++ src/app/(layout-with-navigation)/layout.tsx
@@ -4,6 +4,7 @@ import { useSupportedRegions } from '@/hooks/useSupportedRegions'
 import { PropsWithChildren } from 'react'
 
 import Footer from '@/components/layout/Footer'
+import MetaOGImage from '@/components/misc/MetaOGImage'
 import MobileHeader from './_components/MobileHeader'
 import Navigation from './_components/Navigation'
 
@@ -12,6 +13,8 @@ export default async function PageLayout({ children }: PropsWithChildren) {
 
   return (
     <>
+      <MetaOGImage />
+
       <div className="m-auto flex max-w-7xl justify-start">
         <Navigation />
