+++ src/app/page.tsx
@@ -1,5 +1,7 @@
 import Footer from '@/components/layout/Footer'
 import Logo from '@/components/misc/Logo'
+import MetaOGImage from '@/components/misc/MetaOGImage'
+import MetaOpenGraph from '@/components/misc/MetaOpenGraph'
 import Trans from '@/components/translation/Trans'
 import Main from '@/design-system/layout/Main'
 import Title from '@/design-system/layout/Title'
@@ -13,11 +15,14 @@ import NewsBanner from './_components/NewsBanner'
 import ProfileLink from './_components/ProfileLink'
 import TakeTestLink from './_components/TakeTestLink'
 
+const PAGE_TITLE =
+  "Votre calculateur d'empreinte carbone personnelle - Nos Gestes Climat"
+const PAGE_DESCRIPTION =
+  'Connaissez-vous votre empreinte sur le climat ? Faites le test et découvrez comment réduire votre empreinte carbone sur le climat.'
+
 export const metadata: Metadata = {
-  title:
-    "Votre calculateur d'empreinte carbone personnelle - Nos Gestes Climat",
-  description:
-    'Connaissez-vous votre empreinte sur le climat ? Faites le test et découvrez comment réduire votre empreinte carbone sur le climat.',
+  title: PAGE_TITLE,
+  description: PAGE_DESCRIPTION,
 }
 
 export default async function Landing() {
@@ -28,6 +33,10 @@ export default async function Landing() {
       <header>
         <Logo />
       </header>
+
+      <MetaOpenGraph title={PAGE_TITLE} description={PAGE_DESCRIPTION} />
+
+      <MetaOGImage />
 
       <Main>
         <div className="mx-auto flex flex-col justify-center gap-4 px-4 pb-8 text-center md:mx-auto md:mt-6 md:w-full md:max-w-6xl md:p-10 md:px-8 md:text-left">
