+++ src/helpers/openGraph/generateOGImageURL.ts
@@ -1,0 +1,5 @@
+export function generateOGImageURL(location: string) {
+  return `https://ogimager.osc-fr1.scalingo.io/capture/${encodeURIComponent(
+    location
+  )}/shareImage?timeout=5000`
+}
